+++ src/app/(public)/login/page.tsx
@@ -1,51 +1,16 @@
-'use client';
+import Link from "next/link";
+import LoginForm from "@/components/Forms/LoginForm";
 
-import { FC } from "react";
-import { useRouter } from "next/navigation";
-import { signIn } from "next-auth/react";
-
-import { Form } from "@base-ui-components/react/form";
-import TextField from "@/components/_UI/Forms/TextField";
-import Button from "@/components/_UI/Button";
-
-const LoginPage: FC = () => {
-  const router = useRouter();
-
-  const onSubmit = async (event: { preventDefault: () => void; currentTarget: HTMLFormElement | undefined; }) => {
-    event.preventDefault();
-    const formData = new FormData(event.currentTarget);
-    const values = Object.fromEntries(formData.entries());
-
-    const result = await signIn("credentials", {
-      redirect: false,
-      email: values.email as string,
-      password: values.password as string,
-      callbackUrl: `${window.location.origin}/dashboard`
-    });
-
-    if (result?.error) {
-      console.log("Failed to login:", result.error);
-      // Optionally, set error state here to display error message to user
-    } else {
-      router.push(result?.url || '/dashboard');
-    }
+const LoginPage = () => {
 
   return (
     <>
       <div className="form-container">
         <h1>Login</h1>
-        <Form onSubmit={onSubmit}>
-          <TextField label="Email" name="email" type="email" required />
-          <TextField label="Password" name="password" type="password" required />
-          <div className="btn-container">
-            <Button type="submit" className="submit-button">
-              Login
-            </Button>
-          </div>
-        </Form>
+        <LoginForm />
       </div>
       <div className="center-content">
-        <p>Don't have an account? <a href="/sign-up">Sign Up</a></p>
+        <p>Don't have an account? <Link href="/sign-up">Sign Up</Link></p>
       </div>
     </>
   );
